# core/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "OpenAPI definition",
    "version": "v0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
...
  "paths": {
...
    "/internal/bills/{billId}/transactions": {
      "post": {
        "tags": [
          "internal"
        ],
        "summary": "Выполнить транзакцию по счету у пользователя",
        "description": "Этот метод предназначен для внутреннего сервиса \u0027Кредиты\u0027",
        "operationId": "createTransaction",
        "parameters": [
          {
            "name": "billId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransactionCreationDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "404": {
            "description": "Бизнес ошибка: что-то не найдено",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "400": {
            "description": "Бизнес ошибка или ошибка валидации",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "502": {
            "description": "Ошибка взаимодействия с другими с сервисами",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "500": {
            "description": "Внутренняя ошибка сервиса",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "403": {
            "description": "Недостаточно прав"
          },
          "401": {
            "description": "Не авторизован"
          }
        },
        "security": [
          {
            "Internal Authentication": []
          }
        ]
      }
    },
    "/core/api/v1/internal/bills/{billId}/transactions": {
      "post": {
        "tags": [
          "internal"
        ],
        "summary": "Выполнить транзакцию по счету у пользователя",
        "description": "Этот метод предназначен для внутреннего сервиса \u0027Кредиты\u0027",
        "operationId": "createTransaction_1",
        "parameters": [
          {
            "name": "billId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransactionCreationDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "404": {
            "description": "Бизнес ошибка: что-то не найдено",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "400": {
            "description": "Бизнес ошибка или ошибка валидации",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "502": {
            "description": "Ошибка взаимодействия с другими с сервисами",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "500": {
            "description": "Внутренняя ошибка сервиса",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "403": {
            "description": "Недостаточно прав"
          },
          "401": {
            "description": "Не авторизован"
          }
        },
        "security": [
          {
            "Internal Authentication": []
          }
        ]
      }
    },
...
    "/core/api/v1/internal/bills/{billId}": {
      "get": {
        "tags": [
          "internal"
        ],
        "summary": "Посмотреть информацию о счете пользователя",
        "description": "Этот метод предназначен для внутреннего сервиса \u0027Кредиты\u0027",
        "operationId": "getBill",
        "parameters": [
          {
            "name": "billId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "404": {
            "description": "Бизнес ошибка: что-то не найдено",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "400": {
            "description": "Бизнес ошибка или ошибка валидации",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "502": {
            "description": "Ошибка взаимодействия с другими с сервисами",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "500": {
            "description": "Внутренняя ошибка сервиса",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "403": {
            "description": "Недостаточно прав"
          },
          "401": {
            "description": "Не авторизован"
          }
        },
        "security": [
          {
            "Internal Authentication": []
          }
        ]
      }
    },
    "/internal/bills/{billId}": {
      "get": {
        "tags": [
          "internal"
        ],
        "summary": "Посмотреть информацию о счете пользователя",
        "description": "Этот метод предназначен для внутреннего сервиса \u0027Кредиты\u0027",
        "operationId": "getBill_1",
        "parameters": [
          {
            "name": "billId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "404": {
            "description": "Бизнес ошибка: что-то не найдено",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "400": {
            "description": "Бизнес ошибка или ошибка валидации",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "502": {
            "description": "Ошибка взаимодействия с другими с сервисами",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "500": {
            "description": "Внутренняя ошибка сервиса",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "403": {
            "description": "Недостаточно прав"
          },
          "401": {
            "description": "Не авторизован"
          }
        },
        "security": [
          {
            "Internal Authentication": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "ErrorDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "description": "Код ошибки. Рекомендуется использовать тип ошибки вместо кода.",
            "format": "int32"
          },
          "type": {
            "type": "string",
            "description": "Тип ошибки",
            "enum": [
              "UNKNOWN",
              "BAD_GATEWAY",
              "CANNOT_NEGATIVE_BILL_BALANCE",
              "CLOSING_BILL_WITH_POSITIVE_BALANCE",
              "CLOSING_BILL_WITH_NEGATIVE_BALANCE",
              "TRANSACTION_WITH_ZERO_BALANCE_CHANGE",
              "CLOSING_SYSTEM_BILL",
              "BILL_NOT_FOUND",
              "CLOSED_BILL_OPERATION"
            ]
          },
          "message": {
            "type": "string",
            "description": "Сообщение об ошибке",
            "example": "Сообщение об ошибке"
          }
        }
      },
...
      "TransactionCreationDto": {
        "type": "object",
        "properties": {
          "balanceChange": {
            "type": "integer",
            "description": "Изменение баланса счета в копейках",
            "format": "int64"
          }
        }
      },
...
    }
  }
}